adhere to unit test file naming conventions; separate concerns by moving structs into separate files
