+++ internal/infrastructure/connector/az_blob_connector.go
@@ -17,19 +17,6 @@ import (
 	"github.com/google/uuid"
 )
 
-// BlobConnector is an interface for interacting with Blob storage
-type BlobConnector interface {
-	// UploadFromForm uploads files to a Blob Storage
-	// and returns the metadata for each uploaded byte stream.
-	Upload(ctx context.Context, form *multipart.Form, userId string, encryptionKeyId, signKeyId *string) ([]*blobs.BlobMeta, error)
-
-	// Download retrieves a blob's content by its ID and name, and returns the data as a stream.
-	Download(ctx context.Context, blobId, blobName string) ([]byte, error)
-
-	// Delete deletes a blob from Blob Storage by its ID and Name, and returns any error encountered.
-	Delete(ctx context.Context, blobId, blobName string) error
-}
-
 // AzureBlobConnector is a struct that holds the Azure Blob storage client and implements the BlobConnector interfaces.
 type AzureBlobConnector struct {
 	client        *azblob.Client
