+++ internal/infrastructure/cryptography/ec_test.go
